+++ src/test/java/com/caoccao/javet/swc4j/ast/enums/TestSwc4jAstBinaryOp.java
@@ -16,7 +16,11 @@
 
 package com.caoccao.javet.swc4j.ast.enums;
 
+import com.caoccao.javet.utils.SimpleSet;
 import org.junit.jupiter.api.Test;
+
+import java.util.Set;
+import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -27,10 +31,10 @@ public class TestSwc4jAstBinaryOp {
         assertEquals(Swc4jAstBinaryOp.BitAnd, Swc4jAstBinaryOp.BitOr.getOppositeOperator());
         assertEquals(Swc4jAstBinaryOp.BitOr, Swc4jAstBinaryOp.BitAnd.getOppositeOperator());
         assertEquals(Swc4jAstBinaryOp.Div, Swc4jAstBinaryOp.Mul.getOppositeOperator());
-        assertEquals(Swc4jAstBinaryOp.Gt, Swc4jAstBinaryOp.LtEq.getOppositeOperator());
-        assertEquals(Swc4jAstBinaryOp.GtEq, Swc4jAstBinaryOp.Lt.getOppositeOperator());
         assertEquals(Swc4jAstBinaryOp.EqEq, Swc4jAstBinaryOp.NotEq.getOppositeOperator());
         assertEquals(Swc4jAstBinaryOp.EqEqEq, Swc4jAstBinaryOp.NotEqEq.getOppositeOperator());
+        assertEquals(Swc4jAstBinaryOp.Gt, Swc4jAstBinaryOp.LtEq.getOppositeOperator());
+        assertEquals(Swc4jAstBinaryOp.GtEq, Swc4jAstBinaryOp.Lt.getOppositeOperator());
         assertEquals(Swc4jAstBinaryOp.LShift, Swc4jAstBinaryOp.RShift.getOppositeOperator());
         assertEquals(Swc4jAstBinaryOp.LogicalAnd, Swc4jAstBinaryOp.LogicalOr.getOppositeOperator());
         assertEquals(Swc4jAstBinaryOp.LogicalOr, Swc4jAstBinaryOp.LogicalAnd.getOppositeOperator());
@@ -41,5 +45,35 @@ public class TestSwc4jAstBinaryOp {
         assertEquals(Swc4jAstBinaryOp.NotEqEq, Swc4jAstBinaryOp.EqEqEq.getOppositeOperator());
         assertEquals(Swc4jAstBinaryOp.RShift, Swc4jAstBinaryOp.LShift.getOppositeOperator());
         assertEquals(Swc4jAstBinaryOp.Sub, Swc4jAstBinaryOp.Add.getOppositeOperator());
+    }
+
+    @Test
+    public void testIsLogicalOperator() {
+        Set<Swc4jAstBinaryOp> logicalConditionOperatorSet = SimpleSet.of(
+                Swc4jAstBinaryOp.LogicalAnd,
+                Swc4jAstBinaryOp.LogicalOr);
+        Set<Swc4jAstBinaryOp> logicalCompareOperatorSet = SimpleSet.of(
+                Swc4jAstBinaryOp.EqEq,
+                Swc4jAstBinaryOp.EqEqEq,
+                Swc4jAstBinaryOp.Gt,
+                Swc4jAstBinaryOp.GtEq,
+                Swc4jAstBinaryOp.Lt,
+                Swc4jAstBinaryOp.LtEq,
+                Swc4jAstBinaryOp.NotEq,
+                Swc4jAstBinaryOp.NotEqEq);
+        Stream.of(Swc4jAstBinaryOp.values()).forEach(binaryOp -> {
+            assertEquals(
+                    logicalCompareOperatorSet.contains(binaryOp),
+                    binaryOp.isLogicalCompareOperator(),
+                    binaryOp.name());
+            assertEquals(
+                    logicalConditionOperatorSet.contains(binaryOp),
+                    binaryOp.isLogicalConditionOperator(),
+                    binaryOp.name());
+            assertEquals(
+                    logicalCompareOperatorSet.contains(binaryOp) || logicalConditionOperatorSet.contains(binaryOp),
+                    binaryOp.isLogicalOperator(),
+                    binaryOp.name());
+        });
     }
 }